requirements.txt
# FastAPI and ASGI server
fastapi==0.115.5
uvicorn[standard]==0.34.0
pydantic==2.11.9
pydantic-settings==2.8.0

# Supabase
supabase==2.20.0
postgrest==2.20.0

# Google Gemini AI
google-generativeai==0.8.3

# PDF Processing
pypdf==6.1.1

# Utilities
python-dotenv==1.1.1
python-multipart==0.0.20

# Development
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1
